feat(db): Add new CreateOrderNewList1 function
The new `CreateOrderNewList1` function is added to the `db.ts` file to handle the creation of a new order list one record in the database. This function inserts a new record in the `OrderListOne` table and returns the ID of the newly created record.

This change is necessary to support the new order list one feature in the application.